backend/requirements.txt
sentence-transformers
numpy
spacy 
pinecone
python-dotenv
transformers
torch 
torchvision 
torchaudio
einops
tqdm
pyspellchecker
wordfreq